1.change MLU220_SOC to MLU220EDGE 2.don't log to file as default 3.update live555

+++ framework/core/include/cnstream_pipeline.hpp
@@ -114,7 +114,7 @@ struct LinkStatus {
   std::vector<uint32_t> cache_size;  ///< The size of each queue that is used to cache data between modules.
 };
 
-static constexpr size_t MAX_STREAM_NUM = 64;
+static constexpr uint32_t MAX_STREAM_NUM = 64;
 
 /**
  * @brief ModuleId&StreamIdx manager for pipeline.
